// src/types/rounded.type.ts
export type TRounded =
	| 'rounded-none'
	| 'rounded-xs'
	| 'rounded-sm'
	| 'rounded-md'
	| 'rounded-lg'
	| 'rounded-xl'
	| 'rounded-2xl'
	| 'rounded-3xl'
	| 'rounded-full';

// @start-snippet:: arrRounded
export const arrRounded: TRounded[] = [
	'rounded-none',
	'rounded-xs',
	'rounded-sm',
	'rounded-md',
	'rounded-lg',
	'rounded-xl',
	'rounded-2xl',
	'rounded-3xl',
	'rounded-full',
];
// @end-snippet:: arrRounded